make FoundryCompat version-specific
